requirements.txt
boto3==1.10.39
botocore==1.13.39
Django==2.2
django-cors-headers==3.2.0
django-storages==1.8
djangorestframework==3.10.3
docutils==0.15.2
jmespath==0.9.4
Pillow==6.2.1
python-dateutil==2.8.0
pytz==2019.3
s3transfer==0.2.1
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
